Move ScreenTransitionCancellable from vortex/transitions to vortex/screen
