+++ vortex/src/androidMain/kotlin/io/github/hristogochev/vortex/transitions/AndroidPredictiveBackTransitionCancellable.kt
@@ -6,8 +6,10 @@ import androidx.compose.animation.core.AnimationSpec
 import androidx.compose.animation.core.LinearEasing
 import androidx.compose.animation.core.tween
 import androidx.compose.ui.unit.Density
+import io.github.hristogochev.vortex.screen.ScreenTransitionCancellable
 
-public data class AndroidPredictiveBackTransitionCancellable(val density: Density) : ScreenTransitionCancellable {
+public data class AndroidPredictiveBackTransitionCancellable(val density: Density) :
+    ScreenTransitionCancellable {
     override val zIndex: Float? = -1f
 
     override val cancellableAnimationSpec: AnimationSpec<Float> = tween(
