+++ vortex/src/commonMain/kotlin/io/github/hristogochev/vortex/screen/ScreenTransitionCancellable.kt
@@ -1,7 +1,6 @@
-package io.github.hristogochev.vortex.transitions
+package io.github.hristogochev.vortex.screen
 
 import androidx.compose.animation.core.AnimationSpec
-import io.github.hristogochev.vortex.screen.ScreenTransition
 
 public interface ScreenTransitionCancellable : ScreenTransition {
     public val cancellableAnimationSpec: AnimationSpec<Float>
